Unsupervised_Learning/requirements.txt
Python 3.9.13
numpy==1.24.4
pandas==2.3.2
rdkit==2024.3.6
scikit-learn==1.6.1
umap-learn==0.5.9.post2
matplotlib==3.5.2
matplotlib-inline==0.1.6
typing_extensions==4.14.1
scipy==1.9.1
tqdm==4.64.1    


